arm-none-eabi-gcc: add 6-2017-q2-update release
Create an arm-none-eabi-gcc directory to contain prebuilt GCC cross
toolchains for bare-metal ARM targets for Linux and macOS
hosts. Windows 10 is supported via the Linux subsystem and the Linux
binaries (which are 64 bit executables).

Add NOTICE.txt, which provides information regarding this
redistribution of these toolchains.

The main artifacts we are redistributing are in the linux and mac
subdirectories. For the purposes of GPL compliance, the Corresponding
Sources are included in archive form in a src/ subdirectory, in the
same form the upstream packager provided them.

The GPLed sources are distributed along with the binaries rather than
choosing the written offer compliance route for the purposes of
simplicity. A source offer "must be good for three full years from
your last binary distribution (under GPLv2), or your last binary or
spare part distribution (under GPLv3)" [1], so until we have
infrastructure together to guarantee we can comply, just bundle the
sources with the binary distribution.

[1] https://www.softwarefreedom.org/resources/2008/compliance-guide.html

Signed-off-by: Marti Bolivar <marti.bolivar@linaro.org>